global exception handler

+++ src/main/java/com/todo/todoapp/ToDoService.java
@@ -1,9 +1,14 @@
 package com.todo.todoapp;
 
+import com.todo.todoapp.exception.ToDoNotFoundException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 
 import java.util.List;
 
@@ -16,7 +21,7 @@ public class ToDoService {
         this.repo = repo;
     }
 
-    public Page<ToDo> getAllTasks(@PageableDefault(size = 10,sort = "id") Pageable pageable){
+    public Page<ToDo> getAllTasks(Pageable pageable){
         return repo.findAll(pageable);
     }
 
@@ -28,12 +33,15 @@ public class ToDoService {
     }
 
     public void deleteTask(Integer id){
+        if(!repo.existsById(id)){
+            throw new ToDoNotFoundException(id);
+        }
         repo.deleteById(id);
     }
 
     public ToDo markCompleted(Integer id){
         ToDo todo = repo.findById(id)
-                .orElseThrow(() -> new RuntimeException("Task not found"));
+                .orElseThrow(() -> new ToDoNotFoundException(id));
         todo.setCompleted(true);
         return repo.save(todo);
     }
